+++ interfaces/python/extension/include/osqpworkspacepy.h
@@ -214,7 +214,7 @@
      OSQPSettings *settings = self->workspace->settings;
 
      PyObject *return_dict = Py_BuildValue(
-         "{s:d,s:d,s:i,s:i,s:i,s:d,s:d,s:i,s:d,s:d,s:d, s:d, s:d, s:i, s:i, s:i, s:i}",
+         "{s:d,s:d,s:i,s:i,s:i,s:d,s:d,s:i,s:d,s:d,s:d, s:d, s:d, s:i, s:i, s:i, s:i, s:d}",
          "rho", (double)settings->rho,
          "sigma", (double)settings->sigma,
          "scaling", settings->scaling,
@@ -232,7 +232,7 @@
          "warm_start", settings->warm_start,
          "scaled_termination", settings->scaled_termination,
          "check_termination", settings->check_termination,
-         "time_limit", settings->time_limit);
+         "time_limit", (double)settings->time_limit);
      return return_dict;
  }